Remove unnecessary Moments stuff and fix typos
