+++ packages/moments/src/index.ts
@@ -1,3 +1,3 @@
-export { createMomentInput, FetchMomentsInput } from './types/input';
+export { CreateMomentInput, FetchMomentsInput } from './types/input';
 export { Moment } from './domain/Moment';
 export { MomentsClient } from './MomentsClient';
