+++ packages/providers/src/core/PoapMomentsApi/Types/MediaStatus.ts
@@ -1,0 +1,5 @@
+export enum MediaStatus {
+  INVALID = 'INVALID',
+  PROCESSED = 'PROCESSED',
+  IN_PROCESS = 'IN_PROCESS',
+}
