+++ packages/providers/src/ports/MomentsApiProvider/Types/index.ts
@@ -1,2 +1,2 @@
-export * from './input';
-export * from './response';
+export * from './CreateMomentInput';
+export * from './CreateMomentResponse';
